+++ CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to the BerryCrush IntelliJ Plugin will be documented in this
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.1]
 
-### Added
+### Fixed
+- Fixing missing syntax highlighting for `examples` and tags
 
 ### Changed
+- Highlighting `using` keyword on `call` directive
 
 ## [2.0.0]